add conftest, test_autodetect
