+++ py_src/jupyter_lsp/tests/test_detect.py
@@ -1,0 +1,3 @@
+def test_detect(server):
+    server.initialize(["--LanguageServerApp.autodetect=False"])
+    assert not server.language_servers
