+++ setup.cfg
@@ -44,10 +44,12 @@ jupyter_serverproxy_servers =
     lsp = jupyter_lsp.config:server_process_config
 
 jupyter_lsp_connector_v0 =
-    pyls = jupyter_lsp.connectors:connect_pyls
     javascript_typescript_langserver = jupyter_lsp.connectors:connect_javascript_typescript_langserver
+    pyls = jupyter_lsp.connectors:connect_pyls
+    vscode_css_languageservice = jupyter_lsp.connectors:connect_vscode_css_languageservice
+    vscode_html_languageservice = jupyter_lsp.connectors:connect_vscode_html_languageservice
+    vscode_json_languageservice = jupyter_lsp.connectors:connect_vscode_json_languageservice
     yaml_language_server = jupyter_lsp.connectors:connect_yaml_language_server
-    vscode_json_languageserver = jupyter_lsp.connectors:connect_vscode_json_languageserver
 
 [aliases]
 test = pytest
